+++ src/index.ts
@@ -15,7 +15,7 @@ import {LngLat, LngLatLike} from './geo/lng_lat';
 import {LngLatBounds, LngLatBoundsLike} from './geo/lng_lat_bounds';
 import Point from '@mapbox/point-geometry';
 import {MercatorCoordinate} from './geo/mercator_coordinate';
-import {Evented, ErrorEvent} from './util/evented';
+import {Evented, ErrorEvent, Event} from './util/evented';
 import {config} from './util/config';
 import {rtlMainThreadPluginFactory} from './source/rtl_text_plugin_main_thread';
 import {WorkerPool} from './util/worker_pool';
@@ -191,6 +191,7 @@ export {
     Point,
     MercatorCoordinate,
     Evented,
+    Event,
     AJAXError,
     config,
     CanvasSource,