+++ src/style/style_layer/custom_style_layer.ts
@@ -260,6 +260,8 @@ export function validateCustomStyleLayer(layerObject: CustomLayerInterface) {
     return errors;
 }
 
+export const isCustomStyleLayer = (layer: StyleLayer): layer is CustomStyleLayer => layer.type === 'custom';
+
 export class CustomStyleLayer extends StyleLayer {
 
     implementation: CustomLayerInterface;