+++ src/style/style_layer/hillshade_style_layer.ts
@@ -6,6 +6,8 @@ import {Transitionable, Transitioning, PossiblyEvaluated} from '../properties';
 import type {HillshadePaintProps} from './hillshade_style_layer_properties.g';
 import type {LayerSpecification} from '@maplibre/maplibre-gl-style-spec';
 
+export const isHillshadeStyleLayer = (layer: StyleLayer): layer is HillshadeStyleLayer => layer.type === 'hillshade';
+
 export class HillshadeStyleLayer extends StyleLayer {
     _transitionablePaint: Transitionable<HillshadePaintProps>;
     _transitioningPaint: Transitioning<HillshadePaintProps>;