+++ src/style/style_layer/raster_style_layer.ts
@@ -6,6 +6,8 @@ import {Transitionable, Transitioning, PossiblyEvaluated} from '../properties';
 import type {RasterPaintProps} from './raster_style_layer_properties.g';
 import type {LayerSpecification} from '@maplibre/maplibre-gl-style-spec';
 
+export const isRasterStyleLayer = (layer: StyleLayer): layer is RasterStyleLayer => layer.type === 'raster';
+
 export class RasterStyleLayer extends StyleLayer {
     _transitionablePaint: Transitionable<RasterPaintProps>;
     _transitioningPaint: Transitioning<RasterPaintProps>;