+++ src/shaders/fill_outline.vertex.glsl
@@ -15,7 +15,7 @@ void main() {
     #pragma mapbox: initialize highp vec4 outline_color
     #pragma mapbox: initialize lowp float opacity
 
-    gl_Position = projectTile(a_pos + u_fill_translate);
+    gl_Position = projectTile(a_pos + u_fill_translate, a_pos);
 
     v_pos = (gl_Position.xy / gl_Position.w + 1.0) / 2.0 * u_world;
     #ifdef GLOBE