+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### ✨ Features and improvements
 - Export `Event` class ([#5016](https://github.com/maplibre/maplibre-gl-js/pull/5016))
+- Support Vertical Perspective projection ([#5023](https://github.com/maplibre/maplibre-gl-js/pull/5023))
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
+- Fixes scale control for globe on zoom out ([#4897](https://github.com/maplibre/maplibre-gl-js/pull/4897))
 - _...Add new stuff here..._
 
 ## 5.0.0-pre.6