+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### 🐞 Bug fixes
 - Fixes scale control for globe on zoom out ([#4897](https://github.com/maplibre/maplibre-gl-js/pull/4897))
+- Fixes cooperative gestures displaying the mobile help text when screen width is smaller than 480px on non-touch devices ([#5053](https://github.com/maplibre/maplibre-gl-js/pull/5053))
 - _...Add new stuff here..._
 
 ## 5.0.0-pre.6